move run_manager to own package
