--- pkg/runctl/constants.go
@@ -1,8 +1,0 @@
-package runctl
-
-import (
-	"github.com/SAP/stewardci-core/pkg/k8s"
-)
-
-const runClusterRoleName k8s.RoleName = "steward-run"
-const jfrResultKey string = "jfr-termination-log"
