+++ pkg/runctl/runmgr/constants.go
@@ -1,0 +1,3 @@
+package runmgr
+
+const jfrResultKey string = "jfr-termination-log"
