+++ pkg/runctl/testing/utils.go
@@ -1,31 +1,21 @@
-package runctl
+package testing
 
 import (
 	"fmt"
-	"strings"
 	"testing"
 	"time"
 
-	"gotest.tools/v3/assert/cmp"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-
 	stewardScheme "github.com/SAP/stewardci-core/pkg/client/clientset/versioned/scheme"
+	k8sfake "github.com/SAP/stewardci-core/pkg/k8s/fake"
+	"github.com/SAP/stewardci-core/pkg/runctl/constants"
 	tektonScheme "github.com/SAP/stewardci-core/pkg/tektonclient/clientset/versioned/scheme"
-	"github.com/lithammer/dedent"
+	"gotest.tools/v3/assert/cmp"
+	rbacv1 "k8s.io/api/rbac/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	k8sScheme "k8s.io/client-go/kubernetes/scheme"
 )
-
-// fixIndent removes common leading whitespace from all lines
-// and replaces all tabs by spaces
-func fixIndent(s string) (out string) {
-	const TAB = "   "
-	out = s
-	out = dedent.Dedent(out)
-	out = strings.ReplaceAll(out, "\t", TAB)
-	return
-}
 
 // StewardObjectFromJSON decodes a Steward object from its JSON representation.
 // Panics in case of errors.
@@ -41,7 +31,7 @@ func StewardObjectFromJSON(t *testing.T, doc string) runtime.Object {
 	return obj
 }
 
-// TektonPipelinesObjectFromJSON decodes a Tekton Pipeline object from its JSON
+// TektonObjectFromJSON decodes a Tekton Pipeline object from its JSON
 // representation.
 // Panics in case of errors.
 func TektonObjectFromJSON(t *testing.T, doc string) runtime.Object {
@@ -90,4 +80,9 @@ func TimeEqual(expectedAsRFC3339 string, actual metav1.Time) cmp.Comparison {
 		}
 		return cmp.ResultSuccess
 	}
+}
+
+// FakeClusterRole creates a fake role for testing
+func FakeClusterRole() *rbacv1.ClusterRole {
+	return k8sfake.ClusterRole(string(constants.RunClusterRoleName))
 }
